+++ crates/rendering/src/pipeline/settings.rs
@@ -2,6 +2,7 @@
 use getset::{CopyGetters, Getters};
 #[derive(Getters, CopyGetters)]
 pub struct PipelineSettings {
-    pub shadow_resolution: Option<u16>,
+    // TODO: Generalize
+    pub shadow_resolution: Option<u32>,
     pub shadow_bias: f32,
 }
